(wip) remove all load-file bits, remove all caching stuff, new closure bootstrap approach
