+++ resources/krell_repl.js
@@ -3,7 +3,7 @@ import TcpSocket from "react-native-tcp-socket";
 import { npmDeps } from "./npm_deps.js";
 import { krellNpmDeps } from "./krell_npm_deps.js";
 import { assets } from "./krell_assets.js";
-import AsyncStorage from "@react-native-community/async-storage";
+import "./closure_bootstrap.js";
 
 var CONNECTED = false;
 var RECONNECT_INTERVAL = 3000;
@@ -17,66 +17,7 @@ const KRELL_VERBOSE = $KRELL_VERBOSE;
 const evaluate = eval;
 var libLoadListeners = {};
 var reloadListeners = [];
-var cacheInvalidateListeners = [];
 var pendingLoads_ = [];
-
-// =============================================================================
-// Caching Support
-
-var MEM_CACHE = new Map();
-const CACHE_PREFIX = "krell_cache:";
-
-const isKrellKey = (x) => {
-    return x.indexOf(CACHE_PREFIX) === 0;
-};
-
-const krellPrefix = (x) => {
-    if(isKrellKey(x)) {
-        return x;
-    } else if (typeof x === "string") {
-        return CACHE_PREFIX + x;
-    } else {
-        throw Error("Invalid cache key: " + x);
-    }
-};
-
-const removePrefix = (x) => {
-    if(isKrellKey(x)) {
-        return x.substring(CACHE_PREFIX.length);
-    } else {
-        return x;
-    }
-};
-
-const cacheInit = async () => {
-    let keys = await AsyncStorage.getAllKeys();
-    try {
-        for (let key of keys) {
-            if (isKrellKey(key)) {
-                let path = removePrefix(key);
-                MEM_CACHE.set(path, JSON.parse(await AsyncStorage.getItem(key)));
-            }
-        }
-    } catch(e) {
-        console.error(e);
-        return false;
-    }
-    KRELL_CACHE.ready = true;
-    return true;
-};
-
-const cachePut = (path, entry) => {
-    MEM_CACHE.set(path, entry);
-    if (KRELL_VERBOSE) console.log("CACHE PUT:", path, entry.modified);
-    AsyncStorage.setItem(krellPrefix(path), JSON.stringify(entry))
-        .catch((err) => {
-            console.log("Could not cache path:", path, "error:", err);
-        });
-};
-
-const cacheGet = (path) => {
-    return MEM_CACHE.get(path);
-};
 
 global.CLOSURE_BASE_PATH = "$CLOSURE_BASE_PATH";
 
@@ -84,96 +25,8 @@ function toPath(path) {
     return CLOSURE_BASE_PATH.replace("goog/", "") + path;
 }
 
-// these things are going to change during a single REPL session
-const excludes = {
-    [toPath("goog/base.js")]: true,
-    [toPath("goog/deps.js")]: true,
-    [toPath("cljs_deps.js")]: true,
-    [toPath("krell_repl_deps.js")]: true
-};
-
-const cacheClear = async (all) => {
-    let cacheKeys = MEM_CACHE.keys();
-    MEM_CACHE = new Map();
-    if (all === true) {
-        AsyncStorage.clear();
-    } else {
-        for (let cacheKey of cacheKeys) {
-            if(!excludes[cacheKey]) {
-                if (KRELL_VERBOSE) console.log("CACHE DELETE:", cacheKey);
-                AsyncStorage.removeItem(krellPrefix(cacheKey));
-            }
-        }
-    }
-};
-
-const cacheHas = (path) => {
-    return MEM_CACHE.has(path);
-};
-
-const cacheIsStale = (index) => {
-    if(typeof index == "string") {
-        return index === "invalidate";
-    } else {
-        for (let path of MEM_CACHE.keys()) {
-            let entry = MEM_CACHE.get(path);
-            if (entry) {
-                //console.log(path, entry.modified, index[path]);
-                if (entry.modified < index[path]) {
-                    return true;
-                }
-            } else {
-                return true;
-            }
-        }
-        return false;
-    }
-};
-
-const onKrellCacheInvalidate = (cb) => {
-    cacheInvalidateListeners.push(cb);
-};
-
-const notifyCacheInvalidationListeners = () => {
-    cacheInvalidateListeners.forEach((cb) => cb());
-};
-
-global.KRELL_CACHE = {
-    mem: () => MEM_CACHE,
-    init: cacheInit,
-    put: cachePut,
-    get: cacheGet,
-    has: cacheHas,
-    clear: cacheClear,
-    ready: false
-};
-
-cacheInit();
-
 // =============================================================================
 // REPL Server
-
-var loadFileSocket = null;
-var scheduled = null;
-
-const getSocket = () => {
-    return loadFileSocket;
-};
-
-const loadFile = (socket, path) => {
-    let req = {
-            type: "load-file",
-            value: path
-        },
-        payload = JSON.stringify(req)+"\0";
-    if (!IS_ANDROID) {
-        socket.write(payload);
-    } else {
-        pendingLoads_.push(req);
-        if(scheduled) clearTimeout(scheduled);
-        scheduled = setTimeout(() => flushLoads_(loadFileSocket), 250);
-    }
-};
 
 const exists_ = (obj, xs) => {
     if(typeof xs == "string") {
@@ -201,40 +54,6 @@ const flushLoads_ = (socket) => {
     });
     socket.write(allPending.join(""));
     pendingLoads_ = [];
-};
-
-global.CLOSURE_NO_DEPS = true;
-
-// NOTE: CLOSURE_LOAD_FILE_SYNC not needed as ClojureScript now transpiles
-// offending goog.module files that would need runtime transpiler support
-global.CLOSURE_IMPORT_SCRIPT = function(path, optContents) {
-    if (optContents) {
-        try {
-            evaluate(optContents);
-        } catch (e) {
-            console.error("Could not eval ", path, ":", e);
-            throw e;
-        }
-        return true;
-    } else {
-        var cached = KRELL_CACHE.get(path);
-        if(cached) {
-            try {
-                evaluate(cached.source);
-            } catch (e) {
-                console.error("Could not load ", path, " from cache:", e);
-                throw e;
-            }
-            notifyListeners({value: path});
-        } else {
-            loadFile(loadFileSocket, path, optContents);
-        }
-        return true;
-    }
-};
-
-global.require = function(x) {
-    return npmDeps[x] || krellNpmDeps[x] || assets[x];
 };
 
 const notifyListeners = (request) => {
@@ -288,33 +107,8 @@ const handleMessage = (socket, data) => {
         if(msg.form) {
             ret = evaluate(msg.form);
         }
-        if(req) {
-            if(req.type === "load-file") {
-                let path = req.value;
-                KRELL_CACHE.put(path, {
-                    source: msg.form,
-                    path: path,
-                    modified: req.modified
-                });
-                if (typeof goog !== "undefined") {
-                    goog.debugLoader_.written_[req.value] = true;
-                }
-            }
-            if(req.type === "cache-compare") {
-                if(cacheIsStale(req.index)) {
-                    KRELL_CACHE.clear();
-                    notifyCacheInvalidationListeners();
-                    console.log("Krell cache is stale");
-                } else {
-                    console.log("Krell cache is up-to-date")
-                }
-            }
-        }
     } catch (e) {
         console.error(e);
-        if(req && req.type === "load-file") {
-            console.log("Failed to load file:", req.value);
-        }
         err = e;
     }
 
@@ -345,24 +139,19 @@ const handleMessage = (socket, data) => {
             );
         }
 
+        // TODO: need to move this to new HTTP source file loader
         if (req) {
             notifyListeners(req);
-            // log individual file reloads
-            if(req.reload) {
-                console.log(req);
-            }
-            // all changed files are done reloading, refresh
-            if(req.type === "reload") {
-                notifyReloadListeners();
-            }
         }
     }
 };
 
+global.KRELL_RELOAD = function(nses) {
+    // TODO: notifyReloadListeners() when done
+}
+
 const initSocket = (socket) => {
     var buffer = "";
-    // it doesn't matter which socket we use for loads
-    loadFileSocket = socket;
 
     socket.on("data", data => {
         if (data[data.length - 1] !== 0) {
@@ -411,7 +200,6 @@ tryConnection();
 module.exports = {
     evaluate: evaluate,
     onKrellReload: onKrellReload,
-    onKrellCacheInvalidate: onKrellCacheInvalidate,
     onSourceLoad: onSourceLoad,
     getSocket: getSocket
 };
